// src/screens/StockCheckIn/TransactionTable/propTypes/types.ts
import {Row} from 'component/Table/propTypes/types.ts';

export interface TransactionHistory extends Row {
  orderId: number;
  customerId: number;
  customerName: string;
  grossAmount: number;
  paymentMethods: PaymentMethods;
}

export interface PaymentMethods {
  cash?: number;
  cheque?: number;
  card?: number;
}
export interface PaymentGridProps {
  paymentMethods: PaymentMethods;
}